// src/app/models/funcionario.model.ts
export interface Funcionario {
  funcionarioId?: number;
    identificacion: string;
    nombreFuncionario: string;
    apellidoFuncionario: string;
    celular: string;
    direccion: string;
    email: string;
    estado: string;
    fechaNacimiento: string; 
    cargo: string;
    estadoCivil: string;
    genero: string;
    tipoDocumento: string;
  }